Make dtype of Braid constructor explicit

+++ knpy/braid_vec.py
@@ -43,7 +43,7 @@ class Braid:
         """
         self._braid: BraidNotation
         if isinstance(sigmas, str):
-            self._braid = np.array(knots_in_braid_notation_dict[sigmas][notation_index])
+            self._braid = np.array(knots_in_braid_notation_dict[sigmas][notation_index], dtype=np.int32)
         elif isinstance(sigmas, np.ndarray):
             if copy_sigmas:
                 self._braid = sigmas.copy()
